refactor: drop Files from upload components
