+++ apps/frontend/src/components/ui/servers/UploadDragAndDrop.vue
