+++ apps/frontend/src/components/ui/servers/UploadDropdown.vue
